refactoring

+++ Assets/Game/Scripts/Pathfinding/Editor/WaypointCreator.cs
@@ -1,11 +1,5 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using Cinemachine;
 using UnityEditor;
-using UnityEditor.Rendering;
 using UnityEngine;
-
 public class WaypointCreator : EditorWindow
 {
 #if UNITY_EDITOR
@@ -16,81 +10,37 @@ public class WaypointCreator : EditorWindow
   window.titleContent = new GUIContent("Waypoint Creator");
   window.Show();
  }
- 
  public Transform WaypointParent;
- 
  private void OnGUI()
  {
   var serializedEditorWindow = new SerializedObject(this);
   EditorGUILayout.PropertyField(serializedEditorWindow.FindProperty(nameof(WaypointParent)));
-
-  if (!WaypointParent)
-  {
-   EditorGUILayout.HelpBox("Zieh erstmal ein Pathfinding parent rein", MessageType.Info);
-  }
-  else
-  
-   {
-    DrawButtons();
-   }
-   serializedEditorWindow.ApplyModifiedProperties();
-   
+  if (!WaypointParent) EditorGUILayout.HelpBox("Zieh erstmal ein Pathfinding parent rein", MessageType.Info);
+  else DrawButtons();
+  serializedEditorWindow.ApplyModifiedProperties();
  }
- 
- 
  private void DrawButtons()
  {
-  if (GUILayout.Button("Rename all Points"))
-  {
-   RenameAllWaypoints();
-  }
-  if (GUILayout.Button("Create Point"))
-  {
-   CreateWaypoint();
-  }
-
+  if (GUILayout.Button("Rename all Points")) RenameAllWaypoints();
+  if (GUILayout.Button("Create Point")) CreateWaypoint();
   var selectedWaypoint = GetSelectedWaypoint();
   GUI.enabled = selectedWaypoint;
-
   var waypointName = selectedWaypoint ? $"{selectedWaypoint.name}" : string.Empty;
-
-  if (GUILayout.Button($"[...] <- {waypointName}"))
-  {
-   CreateWaypointBefore(selectedWaypoint);
-  }
-  
-  if (GUILayout.Button($"{waypointName} -> [...]"))
-  {
-   CreateWaypointAfter(selectedWaypoint);
-  }
-  if (GUILayout.Button($"Delete{waypointName}"))
-  {
-  DeleteWaypoint(selectedWaypoint);
-  }
-  
+  if (GUILayout.Button($"[...] <- {waypointName}")) CreateWaypointBefore(selectedWaypoint);
+  if (GUILayout.Button($"{waypointName} -> [...]")) CreateWaypointAfter(selectedWaypoint);
+  if (GUILayout.Button($"Delete{waypointName}")) DeleteWaypoint(selectedWaypoint);
  }
 private void CreateWaypointBefore(Waypoint selectedWaypoint)
   {
    var group = Undo.GetCurrentGroup();
    Undo.RegisterCreatedObjectUndo(selectedWaypoint, "Change Point");
-   if (selectedWaypoint.PreviousPoint != null)
-   {
-    Undo.RegisterCreatedObjectUndo(selectedWaypoint.PreviousPoint, "Change Point");
-   }
-
+   if (selectedWaypoint.PreviousPoint != null) Undo.RegisterCreatedObjectUndo(selectedWaypoint.PreviousPoint, "Change Point");
    var beforeWaypoint = CreateNewWayPoint();
    beforeWaypoint.transform.SetSiblingIndex(selectedWaypoint.transform.GetSiblingIndex());
-
    beforeWaypoint.PreviousPoint = selectedWaypoint.PreviousPoint;
-
-   if (beforeWaypoint.PreviousPoint)
-   {
-    beforeWaypoint.PreviousPoint.NextPoint = beforeWaypoint;
-   }
-
+   if (beforeWaypoint.PreviousPoint) beforeWaypoint.PreviousPoint.NextPoint = beforeWaypoint;
    beforeWaypoint.NextPoint = selectedWaypoint;
    beforeWaypoint.NextPoint.PreviousPoint = beforeWaypoint;
-
    OrientWaypoint(beforeWaypoint, selectedWaypoint);
    Selection.activeGameObject = beforeWaypoint.gameObject;
    RenameAllWaypoints();
@@ -99,27 +49,15 @@ private void CreateWaypointBefore(Waypoint selectedWaypoint)
   }
   private void CreateWaypointAfter(Waypoint selectedWaypoint)
   {
-
    var group = Undo.GetCurrentGroup();
    Undo.RegisterCreatedObjectUndo(selectedWaypoint, "Change Point");
-   if (selectedWaypoint.NextPoint != null)
-   {
-    Undo.RegisterCreatedObjectUndo(selectedWaypoint.NextPoint, "Change Point");
-   }
- 
+   if (selectedWaypoint.NextPoint != null) Undo.RegisterCreatedObjectUndo(selectedWaypoint.NextPoint, "Change Point");
    var afterWaypoint = CreateNewWayPoint();
    afterWaypoint.transform.SetSiblingIndex(selectedWaypoint.transform.GetSiblingIndex()+1);
-
    afterWaypoint.NextPoint = selectedWaypoint.NextPoint;
-
-   if (afterWaypoint.NextPoint)
-   {
-    afterWaypoint.NextPoint.PreviousPoint = afterWaypoint;
-   }
-
+   if (afterWaypoint.NextPoint) afterWaypoint.NextPoint.PreviousPoint = afterWaypoint;
    afterWaypoint.PreviousPoint = selectedWaypoint;
    afterWaypoint.PreviousPoint.NextPoint = afterWaypoint;
-   
    OrientWaypoint(afterWaypoint, selectedWaypoint);
    Selection.activeObject = afterWaypoint.gameObject;
    RenameAllWaypoints();
@@ -134,7 +72,6 @@ private void CreateWaypointBefore(Waypoint selectedWaypoint)
     Undo.RecordObject(selectedWaypoint.PreviousPoint, "Change prev Point");
     selectedWaypoint.PreviousPoint.NextPoint = selectedWaypoint.NextPoint;
    }
-
    if (selectedWaypoint.NextPoint)
    {
     Undo.RecordObject(selectedWaypoint.PreviousPoint, "Change next Point");
@@ -142,11 +79,8 @@ private void CreateWaypointBefore(Waypoint selectedWaypoint)
    }
    Undo.DestroyObjectImmediate(selectedWaypoint.gameObject);
    RenameAllWaypoints();
- 
- 
   }
-
- private Waypoint GetSelectedWaypoint()
+  private Waypoint GetSelectedWaypoint()
  {
   return Selection.activeGameObject ? Selection.activeGameObject.GetComponent<Waypoint>() : null;
  }
@@ -170,29 +104,20 @@ private void CreateWaypointBefore(Waypoint selectedWaypoint)
   {
    waypoint.PreviousPoint =
     WaypointParent.GetChild(waypoint.transform.GetSiblingIndex() - 1).GetComponent<Waypoint>();
-
    waypoint.PreviousPoint.NextPoint = waypoint;
-   
    OrientWaypoint(waypoint, waypoint.PreviousPoint);
   }
   Selection.activeObject = waypoint.gameObject;
-  
  }
-
  private void OrientWaypoint(Waypoint waypoint, Waypoint reference)
  {
-  if (!reference)
-  {
-   return;
-  }
-
+  if (!reference) return;
   var waypointTransform = waypoint.transform;
   var referenceTransform = reference.transform;
   Undo.IncrementCurrentGroup();
   Undo.RecordObject(waypointTransform, $"Change orientation and position of {waypoint}");
   waypointTransform.SetPositionAndRotation(referenceTransform.position, referenceTransform.rotation);
  }
-
  private Waypoint CreateNewWayPoint()
  {
   var waypointGameObject = new GameObject(
@@ -201,7 +126,6 @@ private void CreateWaypointBefore(Waypoint selectedWaypoint)
   waypointGameObject.transform.SetParent(WaypointParent.transform, false);
   Undo.RegisterCreatedObjectUndo(waypointGameObject, "Created new Waypoint");
   return waypointGameObject.GetComponent<Waypoint>();
-  
  }
 #endif
 }
